requirements.txt
Django==4.1.3
django-debug-toolbar==3.7.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
